update app path and URL, stop redirect to intro page
