+++ src/applications/coronavirus-research/containers/ConfirmationPage.jsx
