+++ src/applications/coronavirus-research/containers/IntroductionPage.jsx
@@ -1,0 +1,30 @@
+import React from 'react';
+
+import { focusElement } from 'platform/utilities/ui';
+import FormTitle from 'platform/forms-system/src/js/components/FormTitle';
+import SaveInProgressIntro from 'platform/forms/save-in-progress/SaveInProgressIntro';
+
+class IntroductionPage extends React.Component {
+  componentDidMount() {
+    focusElement('.va-nav-breadcrumbs-list');
+  }
+
+  render() {
+    return (
+      <div className="schemaform-intro">
+        <FormTitle title="Volunteer for COVID-19 research" />
+        <SaveInProgressIntro
+          prefillEnabled={this.props.route.formConfig.prefillEnabled}
+          messages={this.props.route.formConfig.savedFormMessages}
+          pageList={this.props.route.pageList}
+          startText="Volunteer for COVID-19 research"
+        >
+          Please complete the 12345 form to apply for vaccine trial
+          participation.
+        </SaveInProgressIntro>
+      </div>
+    );
+  }
+}
+
+export default IntroductionPage;
