+++ src/applications/coronavirus-research/reducers/index.js
