+++ src/applications/coronavirus-research/tests/00.covid-vaccine-trial.e2e.spec.js
@@ -4,7 +4,7 @@ const manifest = require('../manifest.json');
 
 module.exports = E2eHelpers.createE2eTest(client => {
   client
-    .url(`${E2eHelpers.baseUrl}/covid-research-volunteer`)
+    .url(`${E2eHelpers.baseUrl}/covid-research/volunteer`)
     .waitForElementVisible('body', Timeouts.normal)
     .axeCheck('.main');
