+++ src/applications/coronavirus-research/widgets/HeightWidget.jsx
