--- src/applications/covid-research-volunteer/manifest.json
@@ -1,6 +1,0 @@
-{
-  "appName": "Covid Vaccine Trial",
-  "entryFile": "./app-entry.jsx",
-  "entryName": "covid-research-volunteer",
-  "rootUrl": "/covid-research-volunteer"
-}
