+++ src/applications/registry.json
@@ -827,9 +827,9 @@
     }
   },
   {
-    "appName": "Covid Research - Volunteer",
-    "entryName": "covid-research-volunteer",
-    "rootUrl": "/covid-research-volunteer",
+    "appName": "Coronavirus Research - Volunteer",
+    "entryName": "coronavirus-research",
+    "rootUrl": "/coronavirus-research/volunteer",
     "template": {
       "vagovprod": false,
       "layout": "page-react.html"
